requirements.txt
langchain==0.1.0
langchain-core==0.1.10
langchain-google-genai==0.0.6
google-generativeai==0.3.2
streamlit==1.24.1
python-dotenv==1.0.0
beautifulsoup4==4.13.0
requests==2.31.0 